\name{bee-package}
\alias{bee-package}
\alias{bee}
\docType{package}
\title{
\packageTitle{bee}
}
\description{
\packageDescription{bee}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bee}
\packageIndices{bee}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{bee}

Maintainer: \packageMaintainer{bee}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
