\name{HiveClient-package}
\alias{HiveClient-package}
\alias{HiveClient}
\docType{package}
\title{
\packageTitle{HiveClient}
}
\description{
\packageDescription{HiveClient}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HiveClient}
\packageIndices{HiveClient}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{HiveClient}

Maintainer: \packageMaintainer{HiveClient}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
